# drip-manager/swagger.yaml
swagger: "2.0"
info:
  description: "The Dynamic Real-time infrastructure planner (DRIP) allows application developers to seamlessly plan a customized virtual infrastructure based on application level constraints on QoS and resource budgets, provisioning the virtual infrastructure, deploy application components onto the virtual infrastructure, and start execution on demand using TOSCA."
  version: "3.0.0"
  title: "DRIP"
  contact:
    email: "z.zhao@uva.nl"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
basePath: "/drip-api"
schemes:
- "https"
- "http"
paths:
...
  /planner/plan/{id}:
    get:
      summary: "plan tosca template"
      description: "Returns the ID of the planed topolog template"
      operationId: "planToscaTemplateByID"
      produces:
      - "text/plain"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of topolog template to plan"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "string"
        400:
          description: "Invalid ID supplied"
        404:
          description: "ToscaTemplate not found"
        405:
          description: "Invalid input"          
      security:
      - drip_auth:
        - "write:ToscaTemplate"
        - "read:ToscaTemplate"    
  /planner/ids:
    get:
      summary: "Get all topolog template IDs"
      description: "Returns all IDss "
      operationId: "getPlanToscaTemplateIDs"
      produces:
        - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            type: array
            items:
              type: string
        400:
          description: "Invalid ID supplied"
        404:
          description: "ToscaTemplate not found"
        405:
          description: "Invalid input"          
      security:
      - drip_auth:
        - "write:ToscaTemplate"
        - "read:ToscaTemplate"          
  /planner/{id}:
    get:
      summary: "get the plan tosca template"
      description: "Returns the plan topolog template"
      operationId: "getPlanToscaTemplateByID"
      produces:
      - "text/plain"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of topolog template plan"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "string"
        400:
          description: "Invalid ID supplied"
        404:
          description: "ToscaTemplate not found"
        405:
          description: "Invalid input"          
      security:
      - drip_auth:
        - "write:ToscaTemplate"
        - "read:ToscaTemplate"        
...
  /deployer/deploy/{id}:
    get:
      summary: "deploy the software tosca template"
      description: "Returns the deployment ID"
      operationId: "deployProvisionToscaTemplateByID"
      produces:
      - "text/plain"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of topolog template to deploy"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "string"
        400:
          description: "Invalid ID supplied"
        404:
          description: "ToscaTemplate not found"
        405:
          description: "Invalid input"          
      security:
      - drip_auth:
        - "write:ToscaTemplate"
        - "read:ToscaTemplate"      
  /deployer/{id}:
    get:
      summary: "get the deployment topolog template"
      description: "Returns the deployment topolog template"
      operationId: "getDeployToscaTemplateByID"
      produces:
      - "text/plain"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of topolog template to deploy"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "string"
        400:
          description: "Invalid ID supplied"
        404:
          description: "ToscaTemplate not found"
        405:
          description: "Invalid input"          
      security:
      - drip_auth:
        - "write:ToscaTemplate"
        - "read:ToscaTemplate" 
  /deployer/ids:
    get:
      summary: "Get all topolog template IDs"
      description: "Returns all IDss "
      operationId: "getDeployedToscaTemplateIDs"
      produces:
        - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            type: array
            items:
              type: string
        400:
          description: "Invalid ID supplied"
        404:
          description: "ToscaTemplate not found"
        405:
          description: "Invalid input"          
      security:
      - drip_auth:
        - "write:ToscaTemplate"
        - "read:ToscaTemplate"         
  /credential:
    post:
      summary: "Create credentials"
      description: "Creates credentials"
      operationId: "createCredentials"
      produces:
      - "application/json"
      consumes:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Created user object"
        required: true
        schema:
          $ref: "#/definitions/Credentials"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "string"
        400:
          description: "Invalid ID supplied"
        404:
          description: "ToscaTemplate not found"
        405:
          description: "Invalid input"     
      security:
      - drip_auth:
        - "write:Credentials"
        - "read:Credentials"   
  /credential/ids:
    get:
      summary: "Get all credential IDs"
      description: "Returns all IDss "
      operationId: "getCredentialIDs"
      produces:
        - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            type: array
            items:
              type: string
        400:
          description: "Invalid ID supplied"
        404:
          description: "ToscaTemplate not found"
        405:
          description: "Invalid input"          
      security:
      - drip_auth:
        - "write:ToscaTemplate"
        - "read:ToscaTemplate"            
...
securityDefinitions:
  drip_auth:
    type: "oauth2"
    authorizationUrl: "http://localhost/oauth/dialog"
    flow: "implicit"
    scopes:
      write:ToscaTemplate: "modify topolog template in your account"
      read:ToscaTemplate: "read your topolog template"
      admin:User: "Grants access to admin operations"
      write:Credentials: "modify cloud credentials in your account"
      read:Credentials: "read your cloud credentials"
definitions:
...
  Credentials:
    type: "object"
    properties:
      protocol:
        type: "string"   
      token_type:
        type: "string"        
      token:
        type: "string"        
      keys:
        type: object
        additionalProperties:
          type: string    
      user:
        type: "string"   
      cloud_provider_name:
        type: "string"
...